+++ bocadillo/__init__.py
@@ -1,14 +1,14 @@
-from .applications import API, App
+from .applications import App
 from .errors import HTTPError
 from .middleware import ASGIMiddleware, Middleware
+from .injection import discover_providers, provider, useprovider
 from .recipes import Recipe
+from .request import ClientDisconnect, Request
 from .response import Response
-from .request import ClientDisconnect
-from .request import Request
-from .staticfiles import static
 from .sse import server_event
+from .staticfiles import static
 from .templates import Templates
 from .views import view
 from .websockets import WebSocket, WebSocketDisconnect
 
-__version__ = "0.12.5"
+__version__ = "0.12.6"